requirements-asaphila-bot.txt
requests>=2.28.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
